refactor(api): split api.py into focused modules (closes #4533)
