+++ src/shared/python/humanoid_character_builder/interfaces/__init__.py
@@ -4,10 +4,12 @@ Public interfaces module for humanoid character builder.
 Provides the clean, user-facing API for character building.
 """
 
-from humanoid_character_builder.interfaces.api import (
-    CharacterBuilder,
+from humanoid_character_builder.interfaces.api import CharacterBuilder
+from humanoid_character_builder.interfaces.options import ExportOptions
+from humanoid_character_builder.interfaces.quick import quick_build, quick_urdf
+from humanoid_character_builder.interfaces.results import (
+    BuildErrorCategory,
     CharacterBuildResult,
-    ExportOptions,
     SegmentMeshInfo,
 )
 
@@ -16,4 +18,7 @@ __all__ = [
     "CharacterBuildResult",
     "SegmentMeshInfo",
     "ExportOptions",
+    "BuildErrorCategory",
+    "quick_build",
+    "quick_urdf",
 ]
