+++ tests/unit/tools/humanoid_character_builder/test_api.py
@@ -7,7 +7,8 @@ from pathlib import Path
 
 import pytest
 from humanoid_character_builder.core.body_parameters import BodyParameters
-from humanoid_character_builder.interfaces.api import (
+from humanoid_character_builder.interfaces import (
+    BuildErrorCategory,
     CharacterBuilder,
     ExportOptions,
     quick_build,
